fix(quest): address kilo review feedback on phase 9 quest engine
- journal.hpp, qst_parser.cpp: close `namespace ash` (was `asm`)
- qst_parser.cpp lexer: throw ParseError on stray '+', '-', '!'
  instead of silently returning TokenKind::Eof
- qst_engine.cpp: TimeOfDay now consults hooks_.game_hour() and
  maps hour/day -> daytime (6..18), night -> inverse
- qst_engine.cpp: update() no longer wipes KILL counters when a
  timed stage's remaining seconds hits zero (timeout is owned
  by advance())
- consequence.{hpp,cpp}: ConsequenceRunner now unregisters its
  FlagStore observer in the destructor (prevents dangling-this
  after destruction); add deleted copy ctor/assign
- flag_store.cpp: increment() uses std::stoi per documented
  contract instead of std::stoll
- add trailing newline to all 8 new quest source files

+++ projects/ash/src/quest/flag_store.cpp
@@ -168,7 +168,7 @@ void FlagStore::increment(const std::string& key, std::int64_t delta) {
             } else if constexpr (std::is_same_v<T, std::int64_t>) {
                 cur = x;
             } else {
-                try { cur = static_cast<std::int64_t>(std::stoll(x)); }
+                try { cur = static_cast<std::int64_t>(std::stoi(x)); }
                 catch (...) { cur = 0; }
             }
         }, it->second);
